requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Jinja2==3.1.2
python-dotenv==1.0.0